Cleanup kuttl tests
The existing kuttl tests match the whole Manila CR spec, including
parameters that are not meaninful to make the test pass or not.
This patch removes a lot of fields that are not required in the main
assertion. For example, matching the status is often sufficient to
assert the status of the reconciliation, conditions, and to check if
a resource has been properly deployed.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>